replacing core data description
